+++ CircularBuffer.h
@@ -1,28 +1,29 @@
 #pragma once
 #include <mutex>
+#include <ws2tcpip.h>
 
 class CircularBuffer {
 public:
-	CircularBuffer(size_t bufferSize) : bufferSize(bufferSize), buffer(new char[bufferSize]) {}
+	CircularBuffer(UINT32 bufferSize) : bufferSize(bufferSize), buffer(new char[bufferSize]) {}
 	~CircularBuffer() {
 		delete[] buffer;
 	}
 
     // Write Data
-    bool Write(const char* data, size_t size_) {}
+    bool Write(const char* data, UINT32 size_) {}
 
     // Read Data
-    bool Read(char* readData_, size_t size_) {}
+    bool Read(char* readData_, UINT32 size_) {}
 
-    size_t DataSize() const {}
+	UINT32 DataSize() const {}
 
 private:
 	// 8 bytes
 	char* buffer;
-	const size_t bufferSize;
-	size_t writePos = 0; // Current Write Position
-	size_t readPos = 0;  // Current Read Position
-	size_t currentSize = 0; // Current Buffer Size
+	const UINT32 bufferSize;
+	UINT32 writePos = 0; // Current Write Position
+	UINT32 readPos = 0;  // Current Read Position
+	UINT32 currentSize = 0; // Current Buffer Size
 
     // 80 bytes
     std::mutex bufferMutex;
